Add Traditional Oracle

+++ JDBC/src/main/java/com/example/jdbc/controller/AccountController.java
@@ -1,5 +1,6 @@
 package com.example.jdbc.controller;
 
+import com.example.jdbc.exception.DataNotFoundException;
 import com.example.jdbc.model.Account;
 import com.example.jdbc.service.impl.AccountServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,35 +15,57 @@ import java.util.List;
 @RequestMapping("/account")
 public class AccountController {
     @Autowired
-    @Qualifier("accountServiceImpl")
+//    @Qualifier("accountServiceImpl")
     private AccountServiceImpl accountService;
 
     @GetMapping("")
     public ResponseEntity<List<Account>> findAccounts() {
         List<Account> accounts = accountService.findAccounts();
+        if(accounts.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
         return new ResponseEntity<>(accounts, HttpStatus.OK);
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<Account> findAccountById(@PathVariable int id) {
         Account account = accountService.findAccountById(id);
+        if(account == null) {
+            throw new DataNotFoundException("Data not found");
+        }
         return new ResponseEntity<>(account, HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public int deleteAccount(@PathVariable int id) {
-        return accountService.deleteAccount(id);
+    public ResponseEntity<String> deleteAccount(@PathVariable int id) {
+        Account account = accountService.findAccountById(id);
+        if(account == null) {
+            throw new DataNotFoundException("Data not found");
+        }
+        int executed = accountService.deleteAccount(id);
+        return executed != 0 ? new ResponseEntity<>("Delete account successful", HttpStatus.NO_CONTENT)
+                : new ResponseEntity<>("Delete account failure", HttpStatus.BAD_REQUEST);
+
     }
 
     @PostMapping("")
-    public int addAccount(@RequestBody Account account) {
-        int executedRow = accountService.addAccount(account);
-        return executedRow;
+    public ResponseEntity<String> addAccount(@RequestBody Account account) {
+        int executed = accountService.addAccount(account);
+        return executed != 0 ? new ResponseEntity<>("Add account successful", HttpStatus.CREATED)
+                : new ResponseEntity<>("Add account failure", HttpStatus.BAD_REQUEST);
     }
 
     @PutMapping("")
-    public int editAccount(@RequestBody Account account) {
-        int executedRow = accountService.editAccount(account);
-        return executedRow;
+    public ResponseEntity<String> editAccount(@RequestBody Account acct) {
+        Account account = accountService.findAccountById(acct.getId());
+        if(account == null) {
+            throw new DataNotFoundException("Data not found");
+        }
+        account.setId(acct.getId());
+        account.setEmail(acct.getEmail());
+        account.setPassword(acct.getPassword());
+        int executed = accountService.editAccount(account);
+        return executed != 0 ? new ResponseEntity<>("Edit account successful", HttpStatus.OK)
+                : new ResponseEntity<>("Edit account failure", HttpStatus.BAD_REQUEST);
     }
 }
